subcmd_dump_packed_instructions.cpp

+++ src/subcmd_dump_packed_instructions.cpp
@@ -185,7 +185,7 @@ namespace l
                        offset,
                        next_offset);
             unpack_row(row,bs,ccc);
-            for(uint32_t i = offset; i < next_offset; i++)
+            for(u64 i = offset; i < next_offset; i++)
               {
                 u32 x = bs.read(offset,BITS_PER_BYTE * 4);
                 fmt::print("{:08X} ",x);
